+++ src/DAO/UserDAO.java
@@ -1,7 +1,11 @@
 package DAO;
 
+import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
+import java.util.ArrayList;
+
+import model.User;
 
 public class UserDAO extends BaseDAO {
 	public int getUserIDByEmail(String email) {
@@ -52,7 +56,45 @@ public class UserDAO extends BaseDAO {
 		return null;
 	}
 	
-	public static void main(String[] args) {
-		System.out.println("hello");
+	public ArrayList<User> getAllUser(){ 
+		ArrayList<User> user = new ArrayList<>();
+		try {
+			PreparedStatement ps = connection.prepareStatement("Select * From users WHERE role <> 'admin'");
+			ResultSet rs = ps.executeQuery();
+			while(rs.next()) {
+				int userID = rs.getInt("user_id");
+				String  userName= rs.getString("username");
+				String  email= rs.getString("email");
+				String  photo = rs.getString("photo");			  
+				user.add(new User(userID, userName,email,photo));
+			}
+		} catch (Exception e) {
+		}
+		return user;
 	}
+  public boolean isAdmin(int user_id){
+		String selectAdmin = "Select * From users Where user_id = ?;";
+		try {
+			ResultSet rs = executeQuery(selectAdmin, user_id);
+			if(rs.next()) {
+				String userRole = rs.getString("role");
+				if(userRole.equals("admin")) {
+				return true;
+				}
+			}
+		} catch (SQLException e) {
+         e.printStackTrace();
+		}
+		return false;
+	}
+  public void deleteUserByID(int UserID) {
+		 try {
+			 PreparedStatement ps = connection.prepareStatement("DELETE FROM users WHERE user_id = ?");
+	         ps.setInt(1, UserID);
+	         ps.executeUpdate();
+		} catch (Exception e) {
+			// TODO: handle exception
+		}
+		 
+	 }
 }
